requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.1.31
easyocr==1.7.2
filelock==3.25.2
frozenlist==1.5.0
fsspec==2026.3.0
idna==3.10
ImageIO==2.37.3
Jinja2==3.1.6
lazy-loader==0.5
magic-filter==1.0.12
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.2.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.4
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==12.2.0
propcache==0.3.0
pyclipper==1.4.0
pydantic==2.10.6
pydantic_core==2.27.2
python-bidi==0.6.7
python-dotenv==1.1.0
PyYAML==6.0.3
scikit-image==0.26.0
scipy==1.17.1
shapely==2.1.2
sympy==1.14.0
tifffile==2026.3.3
torch==2.11.0
torchaudio==2.11.0
torchvision==0.26.0
typing_extensions==4.12.2
yarl==1.18.3
